refactor: remove `ZFunction` (#64)
Removes `ZFunction` from the render pipeline completely.

Z Functions were an artifact of my needs at a time before bevy_vello was
merged into Linebender. They are an opinionated component that will
calculate a new Z-index used for rendering for the user, with one of
many functions that would make sense for 2.5D games, 2D platformers,
etc.

Ultimately there's nothing special about Z Functions, they are just a
convenience that can easily exist outside of the rendering pipeline
using a system that changes the Z-value for an asset's transform on
Update, which is why I'm removing them.

This will help to heighten maintainability by reducing quite a bit of
code, and be less confusing for users who would've encountered these.