Pass RenderLayers to VelloCanvas
This PR adds a config option to `VelloPlugin` and `VelloRenderPlugin` that can control the `RenderLayers` that the `VelloCanvas` is on. This allows fully taking advantage of `RenderLayers` as before the canvas was rendering on the default layer, layer 0, which meant any camera that wanted to render it also had to render everything else on the default layer, layer 0.