Pass RenderLayers to VelloCanvas (#71)
This is the fix I talked about in #70 and the PR that would be merged
(the other should be closed)

This PR adds a config option to `VelloPlugin` and `VelloRenderPlugin`
that can control the `RenderLayers` that the `VelloCanvas` is on. This
allows fully taking advantage of `RenderLayers` as before the canvas was
rendering on the default layer, layer 0, which meant any camera that
wanted to render it also had to render everything else on the default
layer, layer 0.

---------

Co-authored-by: Spencer C. Imbleau <spencer@imbleau.com>