refactor: core pipeline changes (#73)
Addresses some friction where the `VelloRenderer` was not a resource and
required Locals temporarily. This was mainly observed in the `cube3d`
example. Likewise, the rendering params such as `use_cpu` and
`antialiasing` were not configurable and the use of `RenderLayers` were
cleaned up after @ChristopherBiscardi brought up some comments after
#71.

@ChristopherBiscardi Please let me know if you still have concerns with
RenderLayers.