Require a VelloView marker for cameras rendering Vello graphics (#85)
It is common for Bevy cameras to mark what graphics features they use
using a marker component, such as bloom or SSR, to name a few built-in
effects that do this. In my game, I have cameras that don't need any
Vello content at all. If I've read the logic correctly, they still pay
the tax of running Vello code.

bevy_vello's `render_frame` system is currently the most demanding
system in my game, and this PR seems to push it down to second place,
although it is quite difficult to measure precisely.

Note that this is a breaking change, since users now have to add the
`VelloView` component to their camera or nothing will render.

---------

Co-authored-by: Spencer C. Imbleau <spencer@imbleau.com>