Calculate text section content size improvements and fixes (#167)
Found a few issues in the calculate text section size systems that this
fixes.

1. Because the systems ran in the `Update` schedule, sometimes when
changing or adding a `Handle<VelloFont>` to a `VelloTextStyle`, the font
would be assigned after calculating the content size.
2. The calculate content size systems previously panicked if the font
was missing, now we gracefully `warn!` instead
3. ~~We did not care about world scale scaling for text sections, I've
added the system and updated the text section queries to select text
sections for the correct space.~~ removed after some reflection, makes
no sense to have it in the plugin